Initial implementation of hiding inactive menu items
We have the need to stop serving a menu item on Dec. 1. I'm taking the
safe road here and instead of deleting the menu item (which at this
point would have unknown effects), I'm adding a date field for setting
inactive_starts_on. This is more flexible than the boolean inactive
field I originally added as it allows us to schedule deactivation.